// source/frontend/src/models/api/generated/ApiGen_Concepts_FormDocumentType.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { ApiGen_Base_BaseAudit } from './ApiGen_Base_BaseAudit';

// LINK: @backend/apimodels/Models/Concepts/FormDocument/FormDocumentTypeModel.cs
export interface ApiGen_Concepts_FormDocumentType extends ApiGen_Base_BaseAudit {
  formTypeCode: string | null;
  documentId: number | null;
  description: string | null;
  displayOrder: number | null;
}